ocr/requirements.txt
fastapi
Pillow
torch
surya-ocr
fastapi
uvicorn
python-multipart
torchvision
aiofiles